[BSE-4473] PyIceberg Write Support (#202)
