+++ BodoSQL/bodosql/tests/test_types/test_tabular_iceberg_catalog.py
@@ -1,13 +1,17 @@
 import os
 import tempfile
+import typing as pt
+from contextlib import contextmanager
 from io import StringIO
 
 import numba
 import pandas as pd
 import pytest
+from pyiceberg.catalog.rest import RestCatalog
 
 import bodo
 import bodosql
+from bodo.mpi4py import MPI
 from bodo.tests.iceberg_database_helpers.utils import (
     get_spark,
 )
@@ -17,9 +21,9 @@ from bodo.tests.user_logging_utils import (
     set_logging_stream,
 )
 from bodo.tests.utils import (
+    _get_dist_arg,
     assert_tables_equal,
     check_func,
-    create_tabular_iceberg_table,
     gen_unique_table_id,
     get_rest_catalog_connection_string,
     pytest_tabular,
@@ -29,6 +33,55 @@ from bodo.utils.utils import run_rank0
 from bodosql.bodosql_types.tabular_catalog import get_tabular_connection
 
 pytestmark = pytest_tabular
+
+
+@contextmanager
+def create_tabular_iceberg_table(
+    df: pd.DataFrame, base_table_name: str, warehouse: str, schema: str, credential: str
+) -> pt.Generator[str, None, None]:
+    """Creates a new Iceberg table in Tabular derived from the base table name
+    and using the DataFrame.
+
+    Returns the name of the table added to Tabular.
+
+    Args:
+        df (pd.DataFrame): DataFrame to insert
+        base_table_name (str): Base string for generating the table name.
+        warehouse (str): Name of the Tabular warehouse
+        schema (str): Name of the Tabular schema
+        credential (str): Credential to authenticate
+
+    Returns:
+        str: The final table name.
+    """
+    comm = MPI.COMM_WORLD
+    iceberg_table_name = None
+    table_written = False
+    try:
+        if bodo.get_rank() == 0:
+            iceberg_table_name = gen_unique_table_id(base_table_name)
+
+        iceberg_table_name = comm.bcast(iceberg_table_name)
+
+        @bodo.jit(distributed=["df"])
+        def write_table(df, table_name, schema, con_str):
+            df.to_sql(table_name, con=con_str, schema=schema, if_exists="replace")
+
+        con_str = f"iceberg+REST://api.tabular.io/ws?credential={credential}&warehouse={warehouse}"
+        write_table(_get_dist_arg(df), iceberg_table_name, schema, con_str)
+        table_written = True
+
+        yield iceberg_table_name
+    finally:
+        if table_written:
+            run_rank0(
+                lambda: (
+                    RestCatalog(
+                        "tabular_catalog",
+                        uri=f"http://api.tabular.io/ws?credential={credential}&warehouse={warehouse}",
+                    ).purge_table(f"{schema}.{iceberg_table_name}")
+                )
+            )()
 
 
 def test_basic_read(memory_leak_check, tabular_catalog):
@@ -275,7 +328,6 @@ def test_tabular_catalog_iceberg_write(
     memory_leak_check,
 ):
     """tests that writing tables works"""
-    import bodo_iceberg_connector as bic
 
     rest_uri, tabular_warehouse, tabular_credential = tabular_connection
     catalog = tabular_catalog
@@ -372,11 +424,11 @@ def test_tabular_catalog_iceberg_write(
         raise e
     finally:
         try:
-            run_rank0(bic.delete_table)(
-                bodo.io.iceberg.format_iceberg_conn(con_str),
-                schema,
-                table_name,
-            )
+            run_rank0(
+                lambda: RestCatalog("tabular_catalog", uri=rest_uri).purge_table(
+                    f"{schema}.{table_name}"
+                )
+            )()
         except Exception:
             if exception_occurred_in_test_body:
                 pass
