+++ bodo/io/arrow_compat.h
@@ -34,15 +34,12 @@ namespace arrow::py {
 
 extern "C++" {
 
-namespace arrow {
-
-namespace py {
+namespace arrow::py {
 int import_pyarrow_wrappers();
 DECLARE_WRAP_FUNCTIONS(dataset, std::shared_ptr<arrow::dataset::Dataset>);
 DECLARE_WRAP_FUNCTIONS(fragment, std::shared_ptr<arrow::dataset::Fragment>);
 DECLARE_WRAP_FUNCTIONS(expression, arrow::compute::Expression);
 DECLARE_WRAP_FUNCTIONS(filesystem, std::shared_ptr<arrow::fs::FileSystem>);
-}  // namespace py
+}  // namespace arrow::py
 
-}  // namespace arrow
 }  // extern "C++"
