+++ bodo/tests/test_iceberg/conftest.py
@@ -1,0 +1,46 @@
+import pytest
+
+from bodo.tests.iceberg_database_helpers.simple_tables import (
+    TABLE_MAP as SIMPLE_TABLES_MAP,
+)
+
+WRITE_TABLES = [
+    "BOOL_BINARY_TABLE",
+    "DT_TSZ_TABLE",
+    "TZ_AWARE_TABLE",
+    "DTYPE_LIST_TABLE",
+    "NUMERIC_TABLE",
+    "STRING_TABLE",
+    "LIST_TABLE",
+    "STRUCT_TABLE",
+    "OPTIONAL_TABLE",
+    # TODO Needs investigation.
+    pytest.param(
+        "MAP_TABLE",
+        marks=pytest.mark.skip(
+            reason="Results in runtime error that's consistent with to_parquet."
+        ),
+    ),
+    pytest.param(
+        "DECIMALS_TABLE",
+        marks=pytest.mark.skip(
+            reason="We don't support custom precisions and scale at the moment."
+        ),
+    ),
+    pytest.param(
+        "DECIMALS_LIST_TABLE",
+        marks=pytest.mark.skip(
+            reason="We don't support custom precisions and scale at the moment."
+        ),
+    ),
+    "DICT_ENCODED_STRING_TABLE",
+]
+
+
+@pytest.fixture(params=WRITE_TABLES)
+def simple_dataframe(request):
+    return (
+        request.param,
+        f"SIMPLE_{request.param}",
+        SIMPLE_TABLES_MAP[f"SIMPLE_{request.param}"][0],
+    )
