+++ bodo/tests/test_iceberg/test_stream_write.py
@@ -1,6 +1,7 @@
 import glob
 import os
 import traceback
+import typing as pt
 
 import pandas as pd
 import pytest
@@ -27,13 +28,12 @@ from bodo.tests.iceberg_database_helpers.utils import (
     create_iceberg_table,
     get_spark,
 )
-from bodo.tests.test_iceberg.test_iceberg import (
+from bodo.tests.test_iceberg.test_iceberg_write import (
     ICEBERG_FIELD_IDS_IN_PQ_SCHEMA_TEST_PARAMS,
     _setup_test_iceberg_field_ids_in_pq_schema,
     _test_file_part,
     _test_file_sorted,
     _verify_pq_schema_in_files,
-    simple_dataframe,  # noqa: F401
 )
 from bodo.tests.utils import (
     _gather_output,
@@ -49,7 +49,12 @@ pytestmark = [pytest.mark.iceberg, pytest.mark.skip(reason="WRITE")]
 
 
 def _write_iceberg_table(
-    df, table_name, conn, db_schema, create_table_meta, if_exists, parallel=True
+    df,
+    table_id: str,
+    conn: str,
+    create_table_meta,
+    mode: pt.Literal["create", "replace", "append"],
+    parallel: bool = True,
 ):
     """helper that writes Iceberg table using Bodo's streaming write"""
 
@@ -57,14 +62,13 @@ def _write_iceberg_table(
     batch_size = 11
 
     @bodo.jit(distributed=(["df"] if parallel else False))
-    def impl(df, table_name, conn, db_schema):
+    def impl(df, table_id, conn):
         writer = iceberg_writer_init(
             -1,
             conn,
-            table_name,
-            db_schema,
+            table_id,
             col_meta,
-            if_exists,
+            mode,
             create_table_meta=create_table_meta,
         )
         all_is_last = False
@@ -84,7 +88,7 @@ def _write_iceberg_table(
             )
             iter_val += 1
 
-    impl((_get_dist_arg(df) if parallel else df), table_name, conn, db_schema)
+    impl((_get_dist_arg(df) if parallel else df), table_id, conn)
 
 
 def init_create_table_meta(
@@ -173,7 +177,7 @@ def check_table_comment(
         .head()
     )
     if table_comments is not None:
-        assert table_cmt[0] == table_comments, (
+        assert table_cmt and table_cmt[0] == table_comments, (
             f'Expected table comment to be "{table_comments}", got "{table_cmt}"'
         )
 
@@ -217,6 +221,7 @@ def test_iceberg_write_basic(
     base_name, table_name, df = simple_dataframe
     db_schema, warehouse_loc = iceberg_database(table_name)
     conn = iceberg_table_conn(table_name, db_schema, warehouse_loc, check_exists=False)
+    table_id = f"{db_schema}.{table_name}"
 
     orig_use_dict_str_type = bodo.hiframes.boxing._use_dict_str_type
     orig_chunk_size = (
@@ -291,9 +296,7 @@ def test_iceberg_write_with_comment_and_properties(
     )
 
     try:
-        _write_iceberg_table(
-            df, table_name, conn, db_schema, create_table_meta, "replace"
-        )
+        _write_iceberg_table(df, table_id, conn, create_table_meta, "replace")
     finally:
         bodo.io.iceberg.stream_iceberg_write.ICEBERG_WRITE_PARQUET_CHUNK_SIZE = (
             orig_chunk_size
@@ -341,6 +344,7 @@ def test_iceberg_write_basic_rep(
     df = SIMPLE_TABLES_MAP["SIMPLE_STRING_TABLE"][0]
     db_schema, warehouse_loc = iceberg_database(table_name)
     conn = iceberg_table_conn(table_name, db_schema, warehouse_loc, check_exists=False)
+    table_id = f"{db_schema}.{table_name}"
 
     # set chunk size to a small number to make sure multiple iterations write files
     bodo.io.iceberg.stream_iceberg_write.ICEBERG_WRITE_PARQUET_CHUNK_SIZE = 300
@@ -348,7 +352,7 @@ def test_iceberg_write_basic_rep(
     create_table_meta = init_create_table_meta()
 
     _write_iceberg_table(
-        df, table_name, conn, db_schema, create_table_meta, "replace", parallel=False
+        df, table_id, conn, create_table_meta, "replace", parallel=False
     )
 
     py_out = bodo.jit()(lambda: pd.read_sql_table(table_name, conn, db_schema))()
@@ -401,6 +405,7 @@ def test_iceberg_write_part_sort(
     bodo.barrier()
     db_schema, warehouse_loc = iceberg_database()
     conn = iceberg_table_conn(table_name, db_schema, warehouse_loc, check_exists=True)
+    table_id = f"{db_schema}.{table_name}"
 
     orig_use_dict_str_type = bodo.hiframes.boxing._use_dict_str_type
     bodo.hiframes.boxing._use_dict_str_type = use_dict_encoding_boxing
@@ -408,12 +413,9 @@ def test_iceberg_write_part_sort(
         bodo.io.iceberg.stream_iceberg_write.ICEBERG_WRITE_PARQUET_CHUNK_SIZE
     )
     bodo.io.iceberg.stream_iceberg_write.ICEBERG_WRITE_PARQUET_CHUNK_SIZE = 300
-    create_table_meta = init_create_table_meta()
 
     try:
-        _write_iceberg_table(
-            df, table_name, conn, db_schema, create_table_meta, "append"
-        )
+        _write_iceberg_table(df, table_id, conn, None, "append")
     finally:
         bodo.hiframes.boxing._use_dict_str_type = orig_use_dict_str_type
         bodo.io.iceberg.stream_iceberg_write.ICEBERG_WRITE_PARQUET_CHUNK_SIZE = (
@@ -467,7 +469,9 @@ def test_iceberg_write_part_sort(
 @pytest.mark.parametrize(
     "base_name,pa_schema", ICEBERG_FIELD_IDS_IN_PQ_SCHEMA_TEST_PARAMS
 )
-@pytest.mark.parametrize("mode", ["create", "replace", "append"])
+@pytest.mark.parametrize("mode", ["create", "append"])
+# TODO: Replace is removed due to inconsistent field ID assignment between PyIceberg
+# and Iceberg Java
 def test_iceberg_field_ids_in_pq_schema(
     base_name,
     pa_schema,
@@ -487,8 +491,7 @@ def test_iceberg_field_ids_in_pq_schema(
     db_schema, warehouse_loc = iceberg_database()
     conn = iceberg_table_conn(table_name, db_schema, warehouse_loc, check_exists=False)
     df, sql_schema = SIMPLE_TABLES_MAP[f"SIMPLE_{base_name}"]
-
-    if_exists = "fail" if mode == "create" else mode
+    table_id = f"{db_schema}.{table_name}"
 
     if base_name == "LIST_TABLE" and mode == "append":
         pytest.skip(
@@ -512,7 +515,7 @@ def test_iceberg_field_ids_in_pq_schema(
     )  # (1KiB)
 
     try:
-        _write_iceberg_table(df, table_name, conn, db_schema, None, if_exists)
+        _write_iceberg_table(df, table_id, conn, None, mode)
     finally:
         bodo.io.iceberg.stream_iceberg_write.ICEBERG_WRITE_PARQUET_CHUNK_SIZE = (
             orig_chunk_size
@@ -568,6 +571,7 @@ def test_iceberg_max_pq_chunksize(
     }
 
     conn = iceberg_table_conn(table_name, db_schema, warehouse_loc, check_exists=False)
+    table_id = f"{db_schema}.{table_name}"
     spark = get_spark()
 
     @run_rank0
@@ -585,7 +589,7 @@ def test_iceberg_max_pq_chunksize(
     total_write_size = sum(df2.memory_usage(index=False))  # = 800000 * 8 bytes
     bodo.barrier()
 
-    _write_iceberg_table(df2, table_name, conn, db_schema, None, if_exists="append")
+    _write_iceberg_table(df2, table_id, conn, None, mode="append")
     bodo.barrier()
     files_after_write = _get_pq_files(warehouse_loc, db_schema, table_name)
 
@@ -599,7 +603,7 @@ def test_iceberg_max_pq_chunksize(
         for file in new_files:
             size_in_bytes = os.stat(file).st_size
             # actual size of files may be much smaller after compression
-            assert size_in_bytes < max_pq_chunksize, (
+            assert size_in_bytes < (max_pq_chunksize * 1.25), (
                 "Found file larger than expected size"
             )
