+++ bodo/tests/test_snowflake_iceberg.py
@@ -1,10 +1,8 @@
-import os
 from io import StringIO
 
 import pandas as pd
 import pyarrow
 import pytest
-from bodo_iceberg_connector.schema import BodoIcebergSchema, get_iceberg_info
 
 import bodo
 from bodo.tests.user_logging_utils import (
@@ -17,7 +15,6 @@ from bodo.tests.utils import (
     get_snowflake_connection_string,
     pytest_mark_one_rank,
     pytest_snowflake,
-    snowflake_cred_env_vars_present,
 )
 
 pytestmark = [pytest.mark.iceberg] + pytest_snowflake
@@ -46,47 +43,6 @@ def test_get_iceberg_schema_snowflake(memory_leak_check):
             pyarrow.field("B", pyarrow.int32()),
         ]
     )
-
-
-@pytest.mark.skip(
-    reason="Current uses get_iceberg_info from the connector which is used for write support. Only read is currently working with PyIceberg"
-)
-@pytest_mark_one_rank
-@pytest.mark.skipif(
-    not snowflake_cred_env_vars_present(),
-    reason="Snowflake environment variables not set",
-)
-def test_get_iceberg_info_snowflake(memory_leak_check):
-    SF_USERNAME = os.environ["SF_USERNAME"]
-    SF_PASSWORD = os.environ["SF_PASSWORD"]
-    info = get_iceberg_info(
-        f"snowflake://bodopartner.us-east-1.snowflakecomputing.com/?warehouse=DEMO_WH&user={SF_USERNAME}&password={SF_PASSWORD}",
-        "TEST_DB.PUBLIC",
-        "TEST_ICEBERG_TABLE",
-    )
-    assert info[0] == 0
-    assert info[1] == "s3://bodo-snowflake-iceberg-test/test_iceberg_table/data"
-    assert info[2] == BodoIcebergSchema(
-        ["VAL", "B"],
-        [
-            pyarrow.field("VAL", pyarrow.bool_(), nullable=False),
-            pyarrow.field("B", pyarrow.int32()),
-        ],
-        [2, 3],
-        [True, False],
-    )
-    assert info[3] == pyarrow.schema(
-        [
-            pyarrow.field("VAL", pyarrow.bool_(), nullable=False),
-            pyarrow.field("B", pyarrow.int32()),
-        ]
-    )
-    assert (
-        info[4]
-        == '{"type":"struct","schema-id":0,"fields":[{"id":2,"name":"VAL","required":true,"type":"boolean"},{"id":3,"name":"B","required":false,"type":"int"}]}'
-    )
-    assert info[5] == []
-    assert info[6] == []
 
 
 def test_basic_read(memory_leak_check):
