+++ iceberg/bodo_iceberg_connector/write.py
@@ -1,6 +1,5 @@
 import base64
 import json
-import typing as pt
 from dataclasses import asdict, dataclass
 from typing import Any
 
@@ -8,26 +7,14 @@ from py4j.protocol import Py4JError
 
 from bodo_iceberg_connector.catalog_conn import (
     gen_file_loc,
-    normalize_data_loc,
     normalize_loc,
     parse_conn_str,
 )
 from bodo_iceberg_connector.puffin import StatisticsFile
 from bodo_iceberg_connector.py4j_support import (
-    convert_dict_to_java,
     convert_list_to_java,
-    get_bodo_arrow_schema_utils_class,
-    get_bodo_iceberg_handler_class,
     get_catalog,
 )
-from bodo_iceberg_connector.schema_helper import (
-    arrow_schema_j2py,
-    arrow_to_iceberg_schema,
-    convert_arrow_schema_to_large_types,
-)
-
-if pt.TYPE_CHECKING:
-    import pyarrow as pa
 
 
 @dataclass
@@ -86,134 +73,6 @@ def process_file_infos(
     return json.dumps(file_infos, cls=BytesEncoder)
 
 
-def get_schema_with_init_field_ids(schema: "pa.Schema") -> "pa.Schema":
-    """
-    Python wrapper around BodoIcebergHandler.getInitSchema.
-    See the docstring of that function for more details about
-    why this is required.
-
-    Args:
-        schema (pa.Schema): Original schema created by Bodo. This
-            must have Iceberg Field IDs embedded in the metadata
-            of the fields.
-
-    Returns:
-        pa.Schema: Same schema, except the field IDs (in the field
-        metadata) would be as if generated by Iceberg Java Library in
-        the final metadata during commit.
-    """
-    handler_class = get_bodo_iceberg_handler_class()
-    bodo_arrow_schema_utils_class = get_bodo_arrow_schema_utils_class()
-    # Get the modified schema in Iceberg Schema format:
-    init_iceberg_schema = handler_class.getInitSchema(arrow_to_iceberg_schema(schema))
-    # Convert the Iceberg schema to a Arrow schema:
-    init_arrow_schema_jvm = bodo_arrow_schema_utils_class.convert(init_iceberg_schema)
-    # Convert the JVM Arrow schema to a PyArrow schema:
-    pyarrow_schema = arrow_schema_j2py(init_arrow_schema_jvm)
-    return convert_arrow_schema_to_large_types(pyarrow_schema)
-
-
-def start_write(
-    conn_str: str,
-    db_name: str,
-    table_name: str,
-    table_loc: str,
-    iceberg_schema_id: int,
-    create_table_info,
-    pa_schema: "pa.Schema",
-    partition_spec: str | None,
-    sort_order: str | None,
-    mode: str,
-):
-    """
-    Start a write action in an Iceberg catalog
-        conn_str: Connection string to catalog
-        db_name: Namespace containing the table written to
-        table_name: Name of table written to
-        table_loc: Warehouse location of data/ path (and files)
-        iceberg_schema_id: Known Schema ID when files were written
-        create_table_info: meta information about table and column comments
-        pa_schema: Arrow Schema of written data. In the create/replace
-            case, make sure to use 'get_schema_with_init_field_ids'
-            to generate the correct field IDs for the fields before
-            writing the parquet files.
-        partition_spec: Iceberg-based partitioning schema of data
-        sort_order: Iceberg-based sorting of data
-        mode: Method of Iceberg write (`create`, `replace`, `append`)
-    returns:
-    txn_id: Transaction ID of the write action
-    table_loc: Updated warehouse location of data/ path (and files)
-    """
-    catalog_type, _ = parse_conn_str(conn_str)
-    handler = get_catalog(conn_str, catalog_type)
-
-    table_comment = None
-    column_comments = None
-    table_properties = None
-
-    if create_table_info is not None:
-        table_comment = create_table_info.table_comment
-        column_comments = create_table_info.column_comments
-        table_properties = create_table_info.table_properties
-
-    additional_properties = {}
-    if table_comment is not None:
-        additional_properties["comment"] = table_comment
-    if table_properties is not None:
-        for key_value in table_properties:
-            additional_properties[key_value[0]] = key_value[1]
-
-    if mode == "create":
-        assert iceberg_schema_id == -1, (
-            "bodo_iceberg_connector Internal Error: Should never create existing table"
-        )
-        try:
-            txn_id = handler.startCreateOrReplaceTable(
-                db_name,
-                table_name,
-                arrow_to_iceberg_schema(pa_schema, column_comments),
-                False,
-                convert_dict_to_java(additional_properties),
-            )
-        except Py4JError as e:
-            print("Error during Iceberg table creation: ", e)
-            return None
-
-    elif mode == "replace":
-        try:
-            txn_id = handler.startCreateOrReplaceTable(
-                db_name,
-                table_name,
-                arrow_to_iceberg_schema(pa_schema, column_comments),
-                True,
-                convert_dict_to_java(additional_properties),
-            )
-        except Py4JError as e:
-            print("Error during Iceberg table replace: ", e)
-            return None
-    else:
-        assert mode == "append", (
-            "bodo_iceberg_connector Internal Error: Unknown write mode. Supported modes: 'create', 'replace', 'append'."
-        )
-        assert iceberg_schema_id is not None
-
-        try:
-            txn_id = handler.startAppendTable(
-                db_name, table_name, convert_dict_to_java(additional_properties)
-            )
-        except Py4JError as e:
-            print("Error during Iceberg table append: ", e)
-            return None
-
-    # Refresh the table location.
-    table_loc = normalize_data_loc(handler.getTransactionTableLocation(txn_id))
-
-    return (
-        txn_id,
-        table_loc,
-    )
-
-
 def commit_write(
     txn_id: int,
     conn_str: str,
@@ -269,27 +128,6 @@ def commit_write(
             print("Error during Iceberg table append: ", e)
             return False
     return True
-
-
-def remove_transaction(
-    transaction_id: int,
-    conn_str: str,
-    db_name: str,
-    table_name: str,
-):
-    """Indicate that a transaction is no longer
-    needed and can be remove from any internal state.
-    This DOES NOT finalize or commit a transaction.
-
-    Args:
-        transaction_id (int): Transaction ID to remove.
-        conn_str (str): Connection string for indexing into our object list.
-        db_name (str): Name of the database for indexing into our object list.
-        table_name (str): Name of the table for indexing into our object list.
-    """
-    catalog_type, _ = parse_conn_str(conn_str)
-    handler = get_catalog(conn_str, catalog_type)
-    handler.removeTransaction(transaction_id)
 
 
 def fetch_puffin_metadata(
